♻️  Refactor: 목데이터 분리, 폴더 구조 개선
