+++ src/components/home/consumptionWelcome.tsx
@@ -1,7 +1,7 @@
 import * as S from '../../styles/home/home.style';
 import { useNavigate } from 'react-router-dom';
 import illust from '../../assets/images/home/illust.png';
-import { mockUser } from '../../mocks/user/mockUser'; // mock data
+import { mockUser } from '../../mocks/home/mockUser'; // mock data
 
 const ConsumptionWelcome = () => {
   const navigate = useNavigate();
