+++ src/mocks/home/mockSpendingData.ts
@@ -1,0 +1,12 @@
+import type { SpendingCategory } from '../../types/home/statistics';
+
+export const mockSpendingData: SpendingCategory[] = [
+  { name: '배달/외식', amount: 80 },
+  { name: '카페', amount: 20 },
+  { name: '고정비', amount: 20 },
+  { name: '패션/쇼핑', amount: 10 },
+  { name: '교육', amount: 30 },
+  { name: '문화생활', amount: 20 },
+  { name: '야구', amount: 70 },
+  { name: '학원', amount: 0 },
+];
